spike/multi-llm/pyproject.toml
[project]
name = "hermes-multi-llm-spike"
version = "0.1.0"
description = "Spike: HERMES multi-LLM adapter pattern"
requires-python = ">=3.11"
dependencies = [
    "google-genai>=1.0.0",
    "anthropic>=0.42.0",
    "pyyaml>=6.0",
]
